Add local whereable trait
